[REFACTOR] :: Refactored css folder structure, added style togheter components folder
