+++ src/components/Carrosel/Carrosel.jsx
@@ -5,7 +5,7 @@ import "swiper/css";
 import "swiper/css/navigation";
 import "swiper/css/pagination";
 
-import '../../styles/Carrosel.css'
+import './Carrosel.css'
 
 const images = [
   "https://blog.emania.com.br/wp-content/uploads/2016/02/direitos-autorais-e-de-imagem.jpg",
