+++ src/components/cardCarousel/CarouselCard.css
@@ -1,0 +1,37 @@
+/* Estiliza o container do Swiper */
+.custom-swiper {
+    max-width: 1600px;
+    padding: 20px 20px 50px 20px; /* Adiciona espaço extra abaixo do Swiper */
+}
+
+/* Estiliza a paginação */
+.custom-swiper .swiper-pagination {
+    margin-top: 20px; /* Adiciona espaço entre os cards e a paginação */
+    bottom: 0; /* Ajusta a posição vertical */
+}
+
+/* Estiliza os bullets da paginação */
+.custom-swiper .swiper-pagination-bullet {
+    background-color: #000; /* Cor dos bullets */
+    opacity: 0.5; /* Transparência */
+    border-radius: 10px;
+    border-radius: 10px;
+    width: 30px;
+
+}
+
+.custom-swiper .swiper-pagination-bullet-active {
+    background-color: #67121B; /* Cor do bullet ativo */
+    opacity: 1; /* Totalmente visível */
+    transition: all 0.5s ease; /* Transição suave ao ativar */
+}
+
+/* Remove estilos padrão do SwiperSlide */
+.swiper-slide {
+    display: flex !important; /* Garante que o layout seja flexível */
+    justify-content: center; /* Centraliza o conteúdo */
+    align-items: center; /* Centraliza verticalmente */
+    height: auto !important; /* Permite que o conteúdo defina a altura */
+    padding: 0 !important; /* Remove qualquer padding padrão */
+    box-sizing: border-box; /* Garante que o tamanho seja calculado corretamente */
+}
