+++ src/components/cards/CardComponent.css
@@ -1,0 +1,45 @@
+.card-button-container{
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    padding: 5px 16px 16px 16px;
+    gap: 5px;
+    position: absolute;
+    bottom: 0;
+    right: 0;
+    left: 0;
+}
+
+.card-button-more{
+    width: 100%;
+    border: none;
+    background-color: transparent;
+    cursor: pointer;
+    font-size: 11px;
+    border-radius: 6px; 
+    transition: all 0.3s ease;
+
+    &:hover{
+        text-decoration: underline;
+    }
+}
+
+.card-content-title{
+    font-size: 16px;
+    line-height: 1;
+}
+
+.card-content-price{
+    font-size: 14px;
+    color: #67121B;
+    margin-bottom: 5px;
+}
+
+.card-content-title,.card-content-price{
+    font-weight: bold;
+}
+
+.card-content-desc{
+    font-size: 11px;
+    line-height: 1.2;
+}
