+++ src/components/cards/CardComponent.jsx
@@ -5,7 +5,7 @@ import CardMedia from '@mui/material/CardMedia';
 
 import WhatsAppButton from './WhatsAppButton';
 
-import '../../styles/CardComponent.css'
+import './CardComponent.css'
 
 
 function CardComponent({props}) {
