+++ src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App'
-import './styles/main.css'
+import './main.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
