--- src/styles/WhatsAppButton.css
@@ -1,20 +1,0 @@
-.whatsapp-button{
-    border: none;
-    cursor: pointer;
-    display: flex;
-    width: 100%;
-    padding: 6px 16px;
-    justify-content: center;
-    align-items: center;
-    gap: 5px;
-    border-radius: 6px;
-    background: #67121B;
-    color: #fff;
-    font-size: 10px;
-    transition: all 0.3s ease;
-
-    &:hover{
-        background-color: #75141e;
-        color: #fff;
-    }
-}
